refactor: remove old MyTeaching code -refs: #1393
