+++ src/client/pages/FeedbackTarget/tabs/InterimFeedback/InterimFeedbackItem.jsx
@@ -9,7 +9,7 @@ import useAuthorizedUser from '../../../../hooks/useAuthorizedUser'
 import useInteractiveMutation from '../../../../hooks/useInteractiveMutation'
 
 import PercentageCell from '../../../CourseSummary/components/PercentageCell'
-import FeedbackResponseChip from '../../../MyTeaching/FeedbackResponseChip'
+import FeedbackResponseChip from '../../../MyTeaching/chips/FeedbackResponseChip'
 import ProtectedRoute from '../../../../components/common/ProtectedRoute'
 
 import { getStartAndEndString } from '../../../../util/getDateRangeString'
