+++ src/client/pages/MyTeaching/chips/FeedbackResponseChip.jsx
@@ -2,8 +2,8 @@
 import React from 'react'
 import { useTranslation } from 'react-i18next'
 import { Chip } from '@mui/material'
-import styles from '../../util/chipStyles'
-import LinkChip from '../../components/common/LinkChip'
+import styles from '../../../util/chipStyles'
+import LinkChip from '../../../components/common/LinkChip'
 
 function getFeedbackUrl(id, interimFeedbackId, feedbackResponseSent, ongoing, continuous) {
   const baseFeedbackUrl = interimFeedbackId ? `/targets/${id}/interim-feedback/${interimFeedbackId}` : `/targets/${id}`
