--- src/client/pages/MyTeaching/index.js
@@ -1,1 +1,0 @@
-export { default } from './MyTeaching'
