+++ src/client/pages/MyTeaching/useMyTeachingTabCounts.js
@@ -1,6 +1,6 @@
 import { useQuery } from 'react-query'
 
-import apiClient from '../../../util/apiClient'
+import apiClient from '../../util/apiClient'
 
 const defaultCacheTime = 900000
